chore(origin-backend): adjust tests to the new Configuration structure
